third_party/blink/web_tests/editing/shadow/selection-of-orphan-shadowroot.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
<script src="../../fast/dom/resources/event-sender-util.js"></script>
<script src="../../fast/dom/resources/shadow-test-driver.js"></script>
</head>
<body>

<p>Nodes of the selection for an orphan shadow root should return null.</p>

<pre id="console"></pre>

<script>
var host = document.createElement('div');
var shadowRoot = host.attachShadow({mode: 'open'});
shadowRoot.innerHTML = "Kotori Otonashi"

var selection = shadowRoot.getSelection();

shouldBe('selection.anchorNode', 'null');
shouldBe('selection.focusNode', 'null');
shouldBe('selection.baseNode', 'null');
shouldBe('selection.extentNode', 'null');

var successfullyParsed = true;
</script>
</body>

</html>
