third_party/blink/web_tests/editing/shadow/selection-of-orphan-shadowroot-expected.txt

Nodes of the selection for an orphan shadow root should return null.

PASS selection.anchorNode is null
PASS selection.focusNode is null
PASS selection.baseNode is null
PASS selection.extentNode is null
PASS successfullyParsed is true

TEST COMPLETE
